Custom key DAO (#275)
Fixes #216 

The DAO generator can now create functions for multi-column keys and generate tests for them. The multi-key approach really stretched the capabilities of the templates, but this should be the complete extent of what we need.

This commit also moves every remaining access through the database object to the DAO (except for the code in the database package, obviously).