fix: ability to skip reading tf_attributes for optional resources (#631)
The `tf_attribute` feature is used to "read back" values from Terraform
in cases where resources have been subsumed, and property values may not
have been stored. This ensures that resources are not overwritten during
subsequent updates. Sometimes the existance of a resource is optional.
For example the Azure brokerpak "csb-azure-mssql-db-failover-group"
offering has an "existing" flag that will use an existing failover group
rather than create one using Terraform. In this case, the reading of
"tf_attribute" needs to be skipped because it will otherwise fail as
there is nothing to read. The new "tf_attribute_skip" field allows
reading of the "tf_attribute" to be optionally skipped depending on the
value of another field.

[#183463447](https://www.pivotaltracker.com/story/show/183463447)