feat: local development experience (#660)
* chore: factor out broker management from brokerpaktestframework

brokerpaktestframework has code that manages starting, stopping and
sending commands to a broker. Very similar things are done in the
integration tests, the example test, and the local CSB experience. This
is a first step that creates a new "testdrive" package which will
contain the common code, and brokerpaktestframework has been modified to
use that package.

Note that this is a technically breaking change because it makes some small changes to the API of brokerpaktestframework - specifically removing the BrokerURL/BrokerUrl methods. But these are not actually called in the AWS, Azure or GCP brokerpaks, so it doesn't actually cause any breakage.

* feat: local development experience

Building on the "csb create-service" experience, this feature enables
the full lifecycle of mimic CF CLI development experience.

The aim is to have commands that look familiar, and allow fast
prototyping and testing of brokerpaks by skipping the need to push the
CSB App to CloudFoundry.

Commands enabled are:

- csb create-service (as before, but can now take parameters)
- csb services (to list created services)
- csb update-service (will allow the no-op update)
- csb delete-service (to clean up)
- csb create-service-key (creates a "binding")
- csb delete-service-key (deletes a "binding")
- csb service-keys (lists "bindings")
- csb marketplace (lists available services and plans)

Co-authored-by: Felisia Martini <fmartini@pivotal.io>