feat: example tests automatically start broker (#714)
Previously, to run the example tests, one had to start a broker in one
terminal, and run the "csb client run-examples" command in another
terminal. This feature builds on the CSB local experience and
automatically starts the broker for you. The new commands are:

- csb examples (to list examples)
- csb run-examples (to run examples)

This feature is experimental for now, and the old functionality still
exists. Assuming the new way works well, we should eventually remove the
old way.