fix: only run database startup when running server (#1012)
An error was discovered where database startup, including database
migrations and failed operation recovery was being run for every command
that accessed the database, including "tf dump". This had the effect of
marking any in-progress operations as failed, even if they were still
really in progress. This has been a problem since the introduction of
in-progress operation recovery in v1.1.0 (commit
6ec2e3e93862efd237375b51d9d691fcaf606c84).

[#187500102](https://www.pivotaltracker.com/story/show/187500102)