feat: enable drain and running on vm (#1097)
* write lockfiles for in flight services

we need a mechanism for the bosh drain lifecycle to know if there are
terraform processes in flight. the lockfiles do just that. they filesystem
approach was favoured over a in memory map to allow an easier interface with
bosh. assuming the csb process crashes while shutting down, recovering the
SIs that were in flight is tricky because the in memory state got lost when
the crash happened.

additionally having the files use the GUIDs for their names allows us to log
the ecact SIs that were not succesfully finished via the drain script.

* add graceful shutdown

we want to enable graceful shutdowns for the csb process so we can
ensure in flight terraform is able to finish before the broker process
exits. this hopes to make the broker more resilient against leaving orphaned
resources in the underlying IaaS.

* move recover in progress to storage package

* update tests to cover shutdown and updated recover in progress

* Changes to address code review

---------

Co-authored-by: Felisia Martini <fmartini@vmware.com>