fix: ParseVcapServices stops searching after the first service offering (#1116)
The for loop on L103 was written to return after the checking the first entry in vcapMap, regardless of whether findMySQLTag found a suitable service or returned an error. If services from multiple service offerings are bound to the application and the service tagged with "mysql" is not in the first offering, the function will not find it. The new test case fails when run against the unmodified code.

To fix this, flatten the map[string][]VcapService into a []VCapService, and pass the flat slice to findMySQLTag.

Co-authored-by: ifindlay-cci <84311346+ifindlay-cci@users.noreply.github.com>